Revert "Add "rigid" slocs that can't unify with each other"
This reverts commit 6ba3cff8abc97e0c20ff264c6a8e5571c03778ce.

This turns out to be a terminally bad idea.  Suppose we have:

  foo(int *x) {
    bar(y);
  }

Then if we treat y as rigid (to avoid parameter aliasing), then x
can't alias with y!  Since we need to introduce some kind of
"uniqueness" constraint to handle this case anyway, it's worth nuking
the whole rigid concept.