don't pull indices for ref variables from integer quals in scalar
For example, suppose we have the qual v > 0 for a ref variable, which
can arise from a call to malloc. This doesn't mean that 1[1] is a legit
index for the resulting ctype; the only predicates we should use to
compute a ref's index should have BLOCK_BEGIN and BLOCK_END in them.